TIP-1145: Fix comments
